log file can skip saving file

+++ FileLogger/Logger.py
@@ -1,13 +1,12 @@
 import os
-import json
-from past.types import basestring
 
 
 class AppLogger:
 
-    def __init__(self, name, clear=False):
+    def __init__(self, name, save_file=True, clear=False):
         self.file_name = name
-        if clear:
+        self.save_file = save_file
+        if self.save_file and clear:
             if os.path.exists(name):
                 os.remove(name)
 
@@ -18,7 +17,8 @@ class AppLogger:
         # else:
         #     text = obj
         print("%s" % obj)
-        f = open(self.file_name, "a")
-        f.write("%s" % obj)
-        f.write("\n")
-        f.close()
+        if self.save_file:
+            f = open(self.file_name, "a")
+            f.write("%s" % obj)
+            f.write("\n")
+            f.close()
